Complete (more or less) universal example
The last thing is not really dev related ... but I'd like to have it
build for prod as well.

There's quite a bit here. I'm reasonably happy with this.

When the client compiles it can take a long time for a refresh, sometimes, a
long time in this case is a few seconds. But it still doesn't feel great
when you can't see that anything is happening. I reckon I might have the
client switch to line by line compiler output during compilation ... for
better feedback.